DSI-8678 moved GetUserServiceAccessDetails to users folder
